Clear narrowing-conversions and enum-size findings (issue #284, part 2/3, batch 2)
bugprone-narrowing-conversions (19): the root cause behind 12 of the 19 was
PieceHelper::Value()/g_iPieceValues being unsigned while every caller uses
piece values in signed score arithmetic (material_score_, MoveHelper::Value,
AIAgent's aspiration window) -- fixed at the source by making both int
instead of patching each call site. The remaining 7: two board-index loops
in defines.h changed their loop variable to int (same class of issue,
smaller scope), AIPerplex.cpp's completion_ratio division made its
int64_t->double narrowing explicit, and TranspositionTable's mate-score
normalize/denormalize made their already-safe (Mate_Threshold + MAX_PLY
fits int16_t) narrowing explicit instead of implicit.

performance-enum-size (9): none of the 9 flagged enums turned out to have a
measurable benefit except GameStates/eSquare, which are GameInfo fields
copied on every DoMove/UndoMove via the 256-entry gameInfoHistory_ -- but
getting real savings there requires the one enum (eSquare) with a
documented, unexplained prior revert. Carved out to #292 rather than
attempted here; suppressed all 9 with NOLINT and specific reasons (4
pointing at #292, 5 confirmed genuinely non-hot: local variables, factory
parameters, or a small non-bulk struct already dominated by a std::string
member).

+++ StratEngine/AIPerplex.cpp
@@ -327,9 +327,11 @@ SearchResult AIPerplex::iterative_deepening(ThreadData& td, int max_depth, Trans
 		metrics.interrupted = ShouldStopSearch(); // Check on time expiry
 		metrics.move_changed = (metrics.current_move != state.last_iteration_move);
 		metrics.score_delta = currentBestScore - state.best_score;
-		metrics.completion_ratio = (state.nodes_at_completed_depth > 0)
-		                               ? static_cast<double>(metrics.nodes_searched) / state.nodes_at_completed_depth
-		                               : 1.0;
+		// node counts realistically never approach 2^53, where int64_t->double would lose precision
+		metrics.completion_ratio =
+		    (state.nodes_at_completed_depth > 0)
+		        ? static_cast<double>(metrics.nodes_searched) / static_cast<double>(state.nodes_at_completed_depth)
+		        : 1.0;
 
 		// Debug logging (detailed diagnostics)
 		log_iteration_eval(metrics, td.pv_table);
